% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble.R
\name{ensemble}
\alias{ensemble}
\title{Evaluate H2O Model(s) Performance}
\usage{
ensemble(
  models,
  training_frame,
  strategy = c("top", "stop"),
  model_selection_criteria = c("auc", "aucpr", "mcc", "f2"),
  max = NULL,
  top_rank = 0.05,
  stop_rank = seq(0.01, 0.99, 0.01),
  stop_rounds = 3,
  reset_stop_rounds = TRUE,
  stop_metric = "auc",
  loaded = TRUE,
  path = NULL,
  seed = -1,
  verbatim = TRUE
)
}
\arguments{
\item{max}{integer. specifies maximum number of models for each criteria to be extracted. the
default value is the \code{"top_rank"} percentage for each model selection
criteria.}
}
\value{
a matrix of F-Measures for different thresholds or the highest F-Measure value
}
\description{
Multiple model performance metrics are computed
}
\examples{

\dontrun{
# initiate h2o server
library(h2o)
h2o.init(ignore_config = TRUE, nthreads = 2, bind_to_localhost = FALSE, insecure = TRUE)

# upload data to h2o cloud
prostate_path <- system.file("extdata", "prostate.csv", package = "h2o")
prostate <- h2o.importFile(path = prostate_path, header = TRUE)

#######################################################
### PREPARE AutoML Grid
#######################################################

# run AutoML to tune various models (GLM, GBM, XGBoost, DRF, DeepLearning) for 30 seconds
y <- "CAPSULE"
prostate[,y] <- as.factor(prostate[,y])  #convert to factor for classification
aml <- h2o.automl(y = y, training_frame = prostate, max_runtime_secs = 30,
                  include_algos=c("DRF","GLM", "XGBoost", "GBM", "DeepLearning"))

#######################################################
### PREPARE H2O Grid (takes a couple of minutes)
#######################################################
grid <- h2o.grid(algorithm = "gbm", y = y, training_frame = prostate,
                hyper_params = list(ntrees = seq(1,500,5)))

#######################################################
### PREPARE ENSEMBLE MODEL
#######################################################

# get the models' IDs from the AutoML grid
ids <- c(h2o.ids(aml), h2o.ids(grid))

# RUN THE ENSEMBLE FOR ALL STRATEGIES
ens <- ensemble(models = ids, training_frame = prostate)

# evaluate the model performance of the best model

}
}
\author{
E. F. Haghish
}
