% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelSelection.R
\name{modelSelection}
\alias{modelSelection}
\title{Evaluate H2O Model(s) Performance}
\usage{
modelSelection(
  eval,
  family = "binary",
  top_rank = 0.01,
  max = round(0.5 * top_rank * nrow(eval)),
  model_selection_criteria = c("auc", "aucpr", "mcc", "f2")
)
}
\arguments{
\item{max}{logical. default is FALSE. if TRUE, instead of providing the F-Measure
for all the thresholds, the highest F-Measure is reported.}

\item{perf}{a h2o object of class \code{"H2OBinomialMetrics"} which is provided
by 'h2o.performance' function.}
}
\value{
a matrix of F-Measures for different thresholds or the highest F-Measure value
}
\description{
Multiple model performance metrics are computed
}
\examples{

\dontrun{
library(h2o)
h2o.init(ignore_config = TRUE, nthreads = 2, bind_to_localhost = FALSE, insecure = TRUE)
prostate_path <- system.file("extdata", "prostate.csv", package = "h2o")
prostate <- h2o.importFile(path = prostate_path, header = TRUE)
y <- "CAPSULE"
prostate[,y] <- as.factor(prostate[,y])  #convert to factor for classification
aml <- h2o.automl(y = y, training_frame = prostate, max_runtime_secs = 30)

# evaluate the model performance
perf <- h2o.performance(aml@leader, xval = TRUE)

# evaluate F-Measure for a Beta = 3
kappa(perf, max = TRUE)
}
}
\author{
E. F. Haghish
}
